refactor: modify local storage hooks directory
