+++ src/client/components/templates/Page/components/NavigationBar/index.tsx
@@ -8,7 +8,7 @@ import { RadioCheckIcon } from 'client/components/atoms/icons/RadioCheckIcon';
 import { SettingsIcon } from 'client/components/atoms/icons/SettingsIcon';
 import { Surface } from 'client/components/atoms/Surface';
 import { Typography } from 'client/components/atoms/Typography';
-import { useLanguageContext } from 'client/store/language/hook';
+import { useLanguageContext } from 'client/store/language/hooks/useLanguageContext';
 import { useOnClickOutside } from 'client/hooks/useOnClickOutside';
 import { componentElevationKey } from 'client/styles/tokens/elevation';
 import { commonStyles, useAppTheme } from 'client/styles/tokens';
@@ -23,7 +23,7 @@ import { TextLink } from 'client/components/molecules/links/TextLink';
 import { BaseButton, BaseButtonRef } from 'client/components/atoms/BaseButton';
 import { useTranslations } from 'client/i18n/hooks/useTranslations';
 import { MENU_BUTTON_ID } from 'client/constants/ids';
-import { useThemeContext } from 'client/store/theme/hook';
+import { useThemeContext } from 'client/store/theme/hook/useThemeContext';
 
 const settingDropdownId = 'setting-dropdown';
 const settingItemClass = 'setting-item';
