+++ src/client/store/language/constants/language.ts
@@ -1,0 +1,5 @@
+export enum Language {
+  En = 'en',
+  Ja = 'ja',
+  Zh = 'zh',
+}
