+++ src/client/store/language/constants/localStorageKey.ts
@@ -1,0 +1,1 @@
+export const LOCAL_STORAGE_LANGUAGE_KEY = 'language';
